Update to get passing build

+++ spec/input/multi_influxdb_spec.rb
@@ -2,32 +2,33 @@ require 'influxdb'
 
 require 'etl/core'
 
-RSpec.describe "influxdb inputs", skip: true do
-  
+RSpec.describe "influxdb inputs" do
+
   let (:dbconfig) {
-    { 
+    {
       :port     => 8086,
       :host     => "127.0.0.1",
       :database => "test"
-    } 
+    }
   }
   let(:idb) { ETL::Input::MultiInfluxdb.new(dbconfig, '', '') }
   let(:ts) { Time.parse('2015-01-10T23:00:50Z').utc } # changing this will break tests
   let(:today) { Time.parse('2015-01-30T23:01:10Z').utc } # changing this will break tests
   let(:option) {
-    { :today => today } 
+    { :today => today }
   }
   let(:series) { 'input_test' }
   let(:container) { 'influx_input_test' }
 
   before(:all) do
-    system("docker run -d -t -p 127.0.0.1:8086:8086 --name multiinflux_input_test influxdb:1.2")
-
-    sleep(0.5) # Give things a second to spin up.
+    if ENV['CIRCLE_CI'] == nil
+      system("docker run -d -t -p 127.0.0.1:8086:8086 --name multiinflux_input_test influxdb:1.2")
+      sleep(0.5) # Give things a second to spin up.
+    end
 
     system("curl -X POST http://127.0.0.1:8086/query --data-urlencode \"q=CREATE DATABASE test\"")
   end
-  
+
   before do
 
     c = idb.conn
@@ -38,7 +39,7 @@ RSpec.describe "influxdb inputs", skip: true do
     for i in 1..1440 do
       h = Hash.new
       h[:series] = series
-      h[:timestamp] = ts.to_i + 20*(i-1) 
+      h[:timestamp] = ts.to_i + 20*(i-1)
       h[:values] = { value: i, n: i*2 }
       color = if i%2 == 1
                 'red'
@@ -58,14 +59,16 @@ RSpec.describe "influxdb inputs", skip: true do
     # 2015-01-10T23:01:10Z	red	  bar	4	2
     # 2015-01-10T23:01:30Z	blue	bar	6	3
   end
-  
+
   after(:all) do
-    system("docker stop multiinflux_input_test")
-    system("docker rm multiinflux_input_test")
+    if ENV['CIRCLE_CI'] == nil
+      system("docker stop multiinflux_input_test")
+      system("docker rm multiinflux_input_test")
+    end
   end
 
   describe 'dummy parameters' do
-  
+
     it 'provides data source name' do
       p = {
         port: 8086,
@@ -78,13 +81,13 @@ RSpec.describe "influxdb inputs", skip: true do
       expect(i.name).to eq("influxdb://test_user@127.0.0.1/metrics")
     end
   end
-  
+
   describe 'test database - first two rows with pagination' do
     let(:testoption) {
       option.merge({:limit => 100})
     }
     let(:midb) { ETL::Input::MultiInfluxdb.new(dbconfig, ["*"], series, testoption) }
-    
+
     it 'returns correct rows' do
       rows = []
       midb.each_row { |row| rows << row }
@@ -94,23 +97,23 @@ RSpec.describe "influxdb inputs", skip: true do
 
       for i in 1..2 do
         h = Hash.new
-        h["time"] = (ts + (i-1)*20).strftime('%FT%TZ') 
+        h["time"] = (ts + (i-1)*20).strftime('%FT%TZ')
         color = if i%2 == 1
                   'red'
                 else
                   'blue'
                 end
         h["color"] = color
-        h["foo"] = "bar" 
-        h["n"] = i*2 
+        h["foo"] = "bar"
+        h["n"] = i*2
         h["value"] = i
         expected.push(h)
       end
-      
+
       expect(rows[0..1]).to eq(expected)
     end
   end
-  
+
   describe 'test database - aggregated by minute and check first three minutes' do
     let(:testoption) {
       option.merge({ :group_by => ["time(1m)"], :where => " time > '2015-01-10T23:00:00Z' and time < '2015-01-10T23:03:00Z' "})
@@ -121,7 +124,7 @@ RSpec.describe "influxdb inputs", skip: true do
       rows = []
       midb.each_row { |row| rows << row }
       expect(midb.rows_processed).to eq(3)
-      
+
       expected = [
         {
           "time" => "2015-01-10T23:00:00Z",
@@ -139,11 +142,11 @@ RSpec.describe "influxdb inputs", skip: true do
           "n" => 3,
         },
       ]
-      
+
       expect(rows).to eq(expected)
     end
   end
-  
+
   describe 'test database - aggregated by color' do
     let(:testoption) {
       option.merge({ :group_by => ["color"], :where => " time >= '2015-01-10T23:00:50Z' and time < '2015-01-30T23:01:10Z' "})
@@ -155,13 +158,13 @@ RSpec.describe "influxdb inputs", skip: true do
     # has given us a time w/o parsing the query, which I want to stay away
     # from for the moment. So we accept that time is in the output. it gets
     # set to the lowest possible time given the range.
-    
+
     it 'returns correct rows' do
       rows = []
       midb.each_row { |row| rows << row }
 
       expect(midb.rows_processed).to eq(2)
-      
+
       expected = [
         {
           "color" => "blue",
@@ -176,7 +179,7 @@ RSpec.describe "influxdb inputs", skip: true do
           "v" => 518400,
         },
       ]
-      
+
       # not sure if influx enforces consistent ordering on these. just sort
       # by color to be safe
       rows.sort! { |a, b| a["color"] <=> b["color"] }
@@ -191,7 +194,7 @@ RSpec.describe "influxdb inputs", skip: true do
         option.merge({ :group_by => ["time(1m)"], :where => " time > '2015-02-10T23:00:00Z' and time < '2015-02-10T23:03:00Z' "})
       }
       let(:midb) { ETL::Input::MultiInfluxdb.new(dbconfig, ["value"], series, testoption) }
-    
+
       it '#rows_processed is zero' do
         expect(midb.rows_processed).to eq(0)
       end
